requirements.txt
python-slugify
colorama
babel
Pillow
PyOpenGraph
unicode
